Fixed boot problems

+++ bestia-messages/src/main/kotlin/net/bestia/messages/attack/AttackUseMessage.kt
@@ -14,7 +14,7 @@ data class AttackUseMessage(
     override val accountId: Long,
 
     @JsonProperty("aid")
-    var attackId: Int,
+    var attackId: Long,
 
     @JsonProperty("x")
     var x: Long,
